backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
python-dotenv==1.0.0
daphne==4.0.0
psycopg2-binary==2.9.9
django-rest-knox==4.2.0